JSON array parsing, float to double

+++ Hero.h
@@ -36,7 +36,7 @@ public:
      * @param damageBonusPerLevel The extra damage added per levelups.
      * @param cooldownMultiplier Multiplier for cooldown on levelup
     */
-    Hero(const std::string &name, unsigned short maxHp, unsigned short damage, float attackCooldown, unsigned short xpPerLevel, unsigned short healthBonusPerLevel, unsigned short damageBonusPerLevel, float cooldownMultiplier) : Unit(name, maxHp, damage, attackCooldown, xpPerLevel, healthBonusPerLevel, damageBonusPerLevel, cooldownMultiplier){};
+    Hero(const std::string &name, unsigned short maxHp, unsigned short damage, double attackCooldown, unsigned short xpPerLevel, unsigned short healthBonusPerLevel, unsigned short damageBonusPerLevel, double cooldownMultiplier) : Unit(name, maxHp, damage, attackCooldown, xpPerLevel, healthBonusPerLevel, damageBonusPerLevel, cooldownMultiplier){};
 
     /**
      * It parse a JSON object (from a JSON file) to a Hero instance.
